refactor: separate main and app logic for better flavor management
