+++ lib/main.dart
@@ -1,31 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:t_store/core/utils/constants/text_strings.dart';
 import 'package:t_store/core/utils/service_locator/service_locator.dart';
-import 'package:t_store/core/utils/theme/theme.dart';
-import 'package:t_store/features/auth/presentation/cubit/on_boarding_cubit.dart';
-import 'package:t_store/features/auth/presentation/views/on_boarding/on_boarding_view.dart';
+import 'package:t_store/t_store.dart';
 
 void main() async {
   await setupServiceLocator();
 
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: TTexts.appName,
-      themeMode: ThemeMode.system,
-      theme: TAppTheme.lightTheme,
-      darkTheme: TAppTheme.darkTheme,
-      debugShowCheckedModeBanner: false,
-      home: BlocProvider(
-        create: (context) => OnBoardingCubit(),
-        child: const OnBoardingView(),
-      ),
-    );
-  }
+  runApp(const TStore());
 }
